backend/requirements.txt
langchain==0.3.18
langchain-core==0.3.28
langchain-openai==0.0.2
langchain-anthropic==0.1.1
langgraph==0.0.20
openai==1.7.2
anthropic==0.8.1
polars==0.20.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
optuna==3.5.0
pyyaml==6.0.1
imbalanced-learn==0.11.0
python-dotenv==1.0.0
